Started Working On Basic Framework For Activation Methods, 5/3/2023
Started Working On Basic Framework For Activation Methods. Still need to figure out how I'm going to be doing all these overall, but hopefully I have somewhat of an idea what I'm going for in the end here, will see tomorrow when I get more of this done and see where it leads to, 5/3/2023.