Seem To Have Save-data On Basic Level Working Now, 5/7/2023
Seem To Have Save-data On Basic Level Working Now. Haven't done any testing with 3D replacement mods and such, but for the sprites/flats atleast the saving and loading part seems to be working fine, will have to see with more testing, but so far so good on something that is often one of the more intimidating parts of making a new mod, imo. Next I'll probably try to get the basic sound-clips to work when taking objects, as well as just generally starting on more of the "bulk" of the mod which will be adding the new items and other stuff involved with taking/interacting with the items in the world and such, 5/7/2023.