Got To Work On Some Of The Tedious Switch-Case Filling, 5/8/2023
Got To Work On Some Of The Tedious Switch-Case Filling. Will continue doing this tomorrow and hopefully once done with that actually maybe start adding custom items in to fill in the mostly place-holder stuff atm. That or actually try to get the "contextual" parts worked out somewhat as well for certain items, will see, 5/8/2023.