Reorganized Code So It's Easier To Work With, 5/9/2023
Reorganized Code So It's Easier To Work With. Now that I have some of the code better organized for the time being, next I work on this I think I'm going to try and start adding in the custom items using most of the framework scripts from Locked Loot Containers, will continue tomorrow, 5/9/2023.