Done With Custom Items Basic Code For Now, 5/12/2023
Done With Custom Items Basic Code For Now. Still may be adding stuff to these custom items later such as "OnUse" things, but that is definitely for later. Right now I'm going to move onto Mod Compatibility and actually adding the various items based on what is being clicked on and such, 5/12/2023.