Changed Various Methods To Allow For Multiple Items, 5/17/2023
Changed Various Methods To Allow For Multiple Items. Now instead of only checking for one item at a time, they all now allow for giving and checking multiple items at one time. So clicking on certain objects can say give more than 1 potion or piece of clothing or whatever at a time, will generally make the code more expandable as well due to this change, I'd imagine. Next time I work on this I'm going to start working on filling in more of the items to be added, all the ones currently not filled in with something that is, 5/17/2023.