Added Methods For Looted Clothing Items, 6/15/2023
Added Methods For Looted Clothing Items. Just getting back on the wagon after nearly an entire month of other stuff distracting me from working on this. So not too bad considering I've been off from this for basically a month and am still a bit rusty. Not bad progress though, next time I'll likely work on doing the same for the weapons and armor items as well, will see, 6/15/2023.