Done With Mod Compatibility Additions For Now, Potions As Well, 6/27/2023
Done With Mod Compatibility Additions For Now, Potions As Well. The potion stuff is still kind of just place-holder compared to what I'd like it to eventually be, but that will likely be for another time post v1.0 of the mod, for right now it is good enough. Tomorrow or next I work on this, will probably finally start on the stuff related to notes and picking up paper and scrolls and stuff, the text there likely being a large bulk of some of the work I'm going to do here, so will just have to see how that turns out, 6/27/2023.