Made Some Progress On Code Structure & Filling Text, 6/29/2023
Made Some Progress On Code Structure & Filling Text. So it does not look like much, but it's very good that I got somewhat of a "plan" seeming to form here in how I may actually do this part, so that is very good progress even though it does not look like much. Also some progress on some of what the next is probably going to look like in some cases, in this case the shopping lists, etc, 6/29/2023.