Slapped Some Stuff Together To See If It Works In-Game, 7/8/2023
Slapped Some Stuff Together To See If It Works In-Game. I'll just have to see tomorrow or whenever the next time I work on this to see if this actually works in-game or not. This is definitely not the way I plan to do all or even any of these necessarily, but more just to see what I can get away with and how I guess, will see, 7/8/2023.