Making This Commit To Save Changes I Will Be Removing, 7/9/2023
Making This Commit To Save Changes I Will Be Removing. So I'm just making this commit to save some changes I made for testing. But during this time I've decided to stop work on this particular text feature for the time being, in the next commit these parts will be removed most likely so I can start work on the other stuff so I can try to release the mod for public use, will see, 7/9/2023.