Added Some Mod Compatibility Stuff & Cleaned Up A Bit, 7/12/2023
Added Some Mod Compatibility Stuff & Cleaned Up A Bit. So added a bit of mod compatibility code for Roleplay Realism: Items mod, atleast for the two added weapons. Also added in a "punishment" check for stealing items inside castles, atleast that's the idea, no idea if the check will actually work or not yet. Finally, did a bit of code clean-up stuff and trimming/commenting out of stuff I'm probably not going to work on until a post v1.0 of this mod update and such, so I can get closer to releasing this for public use. Did a small amount of testing as well, so far so good atleast, but need to do alot more testing before I can say for sure, so will see in the next few days, 7/12/2023.