Did Good Amount More Testing, Fixed Oversights & Bugs, 7/22/2023
Did Good Amount More Testing, Fixed Oversights & Bugs. This was actually the result of a few days, but each day was a pretty small amount of changes so just decided to make the commit today when it was more significant. But yeah, a bunch of smaller stuff changed due to testing and finding some bugs and oversights, so it's good I did do the testing so far, caught a good amount of stuff so far. Still all testing just inside a dungeon of one type, nothing with interior buildings yet, that will be next after testing all of these inside a dungeon, 7/22/2023.